Created Deck.cs and updated accordingly Player.cs
Doubts about Draw() though
